+++ data/nord.jl
@@ -1,0 +1,22 @@
+# the nord theme
+# https://www.nordtheme.com/docs
+# note that the colors in the original scheme are 0-indexed 😱
+
+loadcolorscheme(:nord, [
+    RGB(0.1796875, 0.203125, 0.25),          # nord0
+    RGB(0.23046875, 0.2578125, 0.3203125),   # nord1
+    RGB(0.26171875, 0.296875, 0.3671875),    # nord2
+    RGB(0.296875, 0.3359375, 0.4140625),     # nord3
+    RGB(0.84375, 0.8671875, 0.91015625),     # nord4
+    RGB(0.89453125, 0.91015625, 0.9375),     # nord5
+    RGB(0.921875, 0.93359375, 0.953125),     # nord6
+    RGB(0.55859375, 0.734375, 0.73046875),   # nord7
+    RGB(0.53125, 0.75, 0.8125),              # nord8
+    RGB(0.50390625, 0.62890625, 0.75390625), # nord9
+    RGB(0.3671875, 0.50390625, 0.671875),    # nord10
+    RGB(0.74609375, 0.37890625, 0.4140625),  # nord11
+    RGB(0.8125, 0.52734375, 0.4375),         # nord12
+    RGB(0.91796875, 0.79296875, 0.54296875), # nord13
+    RGB(0.63671875, 0.7421875, 0.546875),    # nord14
+    RGB(0.703125, 0.5546875, 0.67578125),    # nord15
+], "nord", "the Nord theme, Polar Night, Snow Storm, Frost, Aurora")